refactor(mobile): abstract actions row in separate component
