+++ apps/mobile/src/components/ActionsRow/index.ts
@@ -1,0 +1,1 @@
+export { ActionsRow } from './ActionsRow'
