+++ apps/mobile/src/features/ConfirmTx/components/confirmation-views/SwapOrder/SwapOrder.tsx
@@ -1,6 +1,6 @@
 import React, { useMemo } from 'react'
 import { SwapOrderHeader } from './SwapOrderHeader'
-import { View, YStack } from 'tamagui'
+import { YStack } from 'tamagui'
 import { formatSwapOrderItems, formatTwapOrderItems } from './utils'
 import { ListTable } from '../../ListTable'
 import { DataDecoded, MultisigExecutionDetails } from '@safe-global/store/gateway/AUTO_GENERATED/transactions'
@@ -8,16 +8,13 @@ import { OrderTransactionInfo } from '@safe-global/store/gateway/types'
 import { useDefinedActiveSafe } from '@/src/store/hooks/activeSafe'
 import { useAppSelector } from '@/src/store/hooks'
 import { selectChainById } from '@/src/store/chains'
-import { isMultiSendData, isTwapOrderTxInfo } from '@/src/utils/transaction-guards'
+import { isTwapOrderTxInfo } from '@/src/utils/transaction-guards'
 import { isSettingTwapFallbackHandler } from '@safe-global/utils/features/swap/helpers/utils'
 import { TwapFallbackHandlerWarning } from '@/src/features/ConfirmTx/components/confirmation-views/SwapOrder/TwapFallbackHandlerWarning'
 import { Alert } from '@/src/components/Alert'
 import { useRecipientItem } from './hooks'
-import { SafeFontIcon } from '@/src/components/SafeFontIcon'
-import { Badge } from '@/src/components/Badge'
-import { SafeListItem } from '@/src/components/SafeListItem'
-import { useRouter } from 'expo-router'
 import { ParametersButton } from '@/src/features/ConfirmTx/components/ParametersButton'
+import { ActionsRow } from '@/src/components/ActionsRow'
 
 interface SwapOrderProps {
   executionInfo: MultisigExecutionDetails
@@ -28,7 +25,6 @@ interface SwapOrderProps {
 
 export function SwapOrder({ executionInfo, txInfo, decodedData, txId }: SwapOrderProps) {
   const order = txInfo
-  const router = useRouter()
   const isTwapOrder = isTwapOrderTxInfo(order)
 
   const activeSafe = useDefinedActiveSafe()
@@ -45,13 +41,6 @@ export function SwapOrder({ executionInfo, txInfo, decodedData, txId }: SwapOrde
   const recipientItems = useRecipientItem(order)
 
   const showRecipientWarning = order.receiver && order.owner !== order.receiver
-
-  const handleViewActions = () => {
-    router.push({
-      pathname: '/transaction-actions',
-      params: { txId },
-    })
-  }
 
   return (
     <YStack gap="$4">
@@ -73,28 +62,7 @@ export function SwapOrder({ executionInfo, txInfo, decodedData, txId }: SwapOrde
         />
       )}
 
-      {decodedData && isMultiSendData(decodedData) && (
-        <SafeListItem
-          label="Actions"
-          rightNode={
-            <View flexDirection="row" alignItems="center" gap="$2">
-              {decodedData.parameters?.[0]?.valueDecoded && (
-                <Badge
-                  themeName="badge_background_inverted"
-                  content={
-                    Array.isArray(decodedData.parameters[0].valueDecoded)
-                      ? decodedData.parameters[0].valueDecoded.length.toString()
-                      : '1'
-                  }
-                />
-              )}
-
-              <SafeFontIcon name={'chevron-right'} />
-            </View>
-          }
-          onPress={handleViewActions}
-        />
-      )}
+      <ActionsRow txId={txId} decodedData={decodedData} />
     </YStack>
   )
 }
